requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer
click==8.1.3
fastapi==0.89.1
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.1
python-multipart
pytz==2022.7.1
requests
sniffio==1.3.0
starlette==0.22.0
twilio==7.16.2
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.12.1
